requirements.txt
asgiref==3.4.1
Django==3.2.9
django-admin-interface==0.17.3
django-ckeditor==6.2.0
django-colorfield==0.4.5
django-crispy-forms==1.13.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jalali==4.3.0
django-js-asset==1.2.2
jdatetime==3.6.4
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
